+++ exceptions/customExceptions.ts
@@ -1,0 +1,7 @@
+class PasswordsNotMatched extends Error {
+	constructor(code: number, message: string) {
+		super(message);
+	}
+}
+
+export default PasswordsNotMatched;